+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.9.15 - Unreleased
 
+### Fixed
+* Bug in `warp_affine` when the input image has more than 4 dims, has an empty dimension, and `border_value` is an array.
+
 
 ## Version 0.9.14 - Released 2023-03-16